feat(openai): support tagging streamed responses (#8521)
This PR adds support for tagging streamed chat/completion responses from
the OpenAI integration.

Previously we skipped tagging streamed responses as the priority was on
regular responses at the time. However we can now tag as well as include
the responses into the LLMObs span event sent to LLMObs. We do this by
appending each streamed chunk (as it is consumed by the user) to a chunk
list in the completion hook, and then use the collection of chunks to
stitch the response together (and tag it on the span / LLMObs span
event).

The only concerns here are if the user never finishes the stream, then
we won't be able to finish the span. However this is not a new concern
as we had the same behavior prior, now we are at least tagging the
response once finished.

This PR also refactors the llmobs tests in the openai test files to
reuse a util method to construct an LLMObs span event rather than
creating a huge dictionary in every test.

**Note: this PR is nowhere large as the LOC suggests. 90% of the LOC are
in refactored test files + test cassette data.**
Main files to review are:
- ddtrace/contrib/openai/_endpoint_hooks.py
- ddtrace/contrib/openai/utils.py
- ddtrace/llmobs/_integrations/bedrock.py
- ddtrace/llmobs/_integrations/openai.py

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.
- [x] If change touches code that signs or publishes builds or packages,
or handles credentials of any kind, I've requested a review from
`@DataDog/security-design-and-guidance`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [X] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)