chore(di): add probe location kind
We add support for the new probe location kind field. This is used to
replace the old logic of inferring the probe location kind on the basis
of the property contents of the probe location attribute. Instead, we
make the logic more robust by basing it on the value of the new extra
field.