fix(internal): chained namespace loader
We fix the implemention of the support for namespace module imports that
caused issues with some standard library modules, such as `importlib`.
We make sure that the internal module attribute initialisation can
create the correct namespace loader, and then we make sure that we chain
it with the custom loader used to trigger import hooks.

Addresses #8003