chore(ci_visibility): experimental CI Visibility manual testing API (#8711)
Introducing an API that will eventually supplant a lot of the code that
currently lives in the `pytest` and `unittest` integrations.

It is currently in an incomplete state to minimize the size of the
introductory change and allow for smaller iterative updates. Feature
parity with existing integrations is a non-goal. Some of the features
intended to be supported by some of the flags introduced by the API are
also unimplemented.

### `ddtrace.ext.ci_visibility.api`

This is the user-facing API that contains the various methods,
structures (eg: dataclasses, named tuples, etc) to be used when
interacting with the CIVisibility service. It is all in one file to give
a single-point-of-reference, rather than having to look across multiple
modules for information.

It is meant to be used in a stateless way (without users needing to keep
track of objects representing test items). The API emits events (eg:
"session start", "test discovered" , "test started", "test finished") on
a best-effort basis. Classes and classmethods are only used for
namespacing purposes.

### `ddtrace.internal.ci_visibility.api`

Introduces classes to represent CI Visibility items, and stores them on
the CIVisibility singleton instance's `_session_data` attribute. The
code allows for concurrent sessions, though that is not an expected or
supported behavior.

To support non-linear progression through modules and suites, spans are
started with their parent spans manually specified, and without the core
execution contexts API (subject to change).

### Other changes

Several setter/getter classmethods are introduced in the `CIVsibility`
class to respect the use of `_instance` as a private attribute.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.
- [x] If change touches code that signs or publishes builds or packages,
or handles credentials of any kind, I've requested a review from
`@DataDog/security-design-and-guidance`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)