feat(tracing): lazy sampling (#8308)
This PR makes sampling lazy, effectively making our extended sampling
API that can take resource and operation name far more useful, since we
will now always be sampling after those tags have been added. Currently,
often resource is added afterwards and operation name can change,
therefore this PR essentially finishes the extension.

To make this PR simpler for reviewing, changes in this PR can be
categorized as such:

1. Do not sample in `_start_span`

2. Sample right before propagation by modifying all of the
`HTTPPropagator.inject()` calls to take a tracer and span so we can
sample there right before the sampling decision is propagated.

3. Sample right before encoding using `TraceSamplingProcessor`

4. Sample right before  `_DBM_Propagator.inject()`

5. Sample right before `os.fork`


**Design Decisions:**

1. _Why are we running `context._update_tags` when we sample?_
We need the tags and metrics from the context on the root_span when we
sample. This unfortunately leads to us running update_tags twice, once
when we sample, and once in the TraceTagsProcessor right before we
encode the trace chunk._

The alternative was pulling out the `context._meta` and
`context._metrics` values to combine them with the span's metrics and
tags values in `sampling_rule.py::tags_match` or running sampling on
those values in addition to the span's values to come up with a match.
Both of these felt messier, and technically the possibility exists that
the same tag key could exist on the context and the span, in which case
the current span key should take precedence (as is the current behavior
with context._update_tags). This could obviously be implemented with a
`continue` in the for loop evaluating tag rules against spans, however
it adds to the mess. I'm still kind of on the fence about this choice
and am happy to discuss.

I've added the update in sample, because there are multiple other places
where we need to make sure that we update, before sampling, e.g. before
propagation http, before propagation dbm. Therefore I didn't want to
make sure we always ran context.update_tags after every egress point
correctly before we sample.

2. _It's kind of messy to modify the `HttpPropagator.inject` signature
to take a span._

No doubt, but I couldn't find a way around this besides either wrapping
inject, with a method that samples before running it, or having to
sample every place we call `HttpPropagator.inject`. It seemed safer to
me to call sampling inside of it. However it isn't great, because if
customers are using inject() int their code, don't pass in a span, then
sampling is not run. I will add a rn about that.
Another option was to always run `ddtrace.tracer.current_root_span()` to
grab the span, which we do as a back up if inject is called without a
span kwarg value. However, Emmett and I decided that the ability to pass
in an explicit span if possible should be included be the norm in our
use.
I'm open to any suggestions regarding this.

3. _Ok so you instrumented os.fork to sample before we fork, but what
about all of the other ways you can change execution context? E.g.
spawn, pOpen, etc.? Can't that lead to differences in
context.sampling_priority between the parent process's spans and the
child processes spans?_

Yes, it can. I'm not sure I can add instrumentation to all of the ways
you can change execution context in this PR, however I'd be happy to
hear advice on how to add it to any easy to reach places. I've also
added to [the
documentation](https://github.com/DataDog/dd-trace-py/pull/8308/files#diff-151a99faebb7c20650eb815a2821b6613090470c2873994c64334fc9190ee1b0R84-R154)
that you should sample manually before changing execution context.

4. _Follow up, to number 3: Does this break gunicorn or gevent since for
execution context switching I've only changed os.fork to sample before
the context change?_

No it does not, we don’t actually ever start traces in the parent
process, the request is just routed to the worker, and the trace starts
there. In the event that middleware is added to a gunicorn + gevent
application the middleware sits between the parent process routing the
request and the worker, so sampling propagation should work fine there
as well, as long as the middleware and the application logic are
executing in the same execution context.

5. _What about when customers have an egress point that's not
automatically instrumented or they don't use `HttpPropagator.inject`
for? Won't that break sampling?_

Yes, they'll need to sample manually before the injection. I've added
that here in [the
docs](https://github.com/DataDog/dd-trace-py/pull/8308/files#diff-151a99faebb7c20650eb815a2821b6613090470c2873994c64334fc9190ee1b0R241-R243).

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.
- [x] If change touches code that signs or publishes builds or packages,
or handles credentials of any kind, I've requested a review from
`@DataDog/security-design-and-guidance`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Tahir H. Butt <tahir.butt@datadoghq.com>
Co-authored-by: Emmett Butler <723615+emmettbutler@users.noreply.github.com>
Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>