fix(setuptools): use guess-next-dev schema
A backport (#8367) changed our setuptools_scm schema from
`guess-next-dev` to `release-branch-semver` for the 2.5 and 2.6
releases; but this change was never put into mainline.  This change
fixes the issue.

Prior to this change, setuptools_scm reported the version for patch
releases as the next minor release.  Afterwards, versions are patch
releases (as expected).