chore(profiling): add non-running tests for stack v2 (#8576)
This PR creates a harness for independently building and running the
stack v2 code. I did it this way for a few reasons

* Don't want to add unnecessary overhead to our normal test process
* Want to run the build for these native components under multiple
static analysis tools
* Want to run some of the tests for these components under sanitizers?
* But also want to do exhaustive testing of the interfaces, since we've
had a few regressions here

This PR includes both the tests and the fixups they demand. A separate
PR will add a CI job to execute these (or will add it to a pre-existing
job, TBD).

There IS some dead code in here, since we do provide the recipe for
getting and running Infer. During testing, I found that I couldn't
reliably get infer to find stdlib headers, which completely broke the
quality of its output, so it won't actually be used. At the same time, I
think we might want to use it someday, and the setup is a little fiddly,
so I'm including it here.

I'm recommending a backport to 2.7 for this because the latest 2.7
releases make this feature available, but the implementation is subtly
(and sometimes not-so-subtly) broken without the fixups here. Since this
feature is so isolated (I'm the only one backporting anything related!),
this feels safe...ish.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.
- [x] If change touches code that signs or publishes builds or packages,
or handles credentials of any kind, I've requested a review from
`@DataDog/security-design-and-guidance`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: sanchda <sanchda@users.noreply.github.com>
Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>