fix(pymongo): fix client validation issue
Prior to this change, there were some scenarios in which pymongo would
call `validate_session` where it asserted that the client which created
the session was the client attempting to make a call with that session.

Since we are using ObjectProxy to wrap the MongoClient, we're actually
forwarding requests to a separate client.  In this case, we run into
trouble in the validation situation where TracedMongoClient forwards
requests to MongoClient, which then creates a session and sets the
session create as itself (thus bypassing the proxy).

After this fix, we intercept calls to validate_session and make ensure
we're comoparing apples with apples.