feat(tracer): isolated tracer flare logic (#8961)
## Overview
This PR adds the tracer flare logic to dd-trace-py without wiring it up
to the RC subscriber. Therefore, these changes will not be visible on
the customer end yet. There will be a follow-up PR to wire this up.

A tracer flare is a tar file containing tracer logs from the last 5
minutes, as well as a JSON file of the current tracer configurations.
This pair of files will be generated per tracer instance that is
connected to the agent, and the tar containing all of these files will
be sent to a Zendesk ticket via an endpoint on the ddagent. Users can
trigger the tracer flare the same way they trigger agent flares. See
Agent Flare documentation
[here](https://docs.datadoghq.com/agent/troubleshooting/send_a_flare/?tab=agentv6v7).

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)