fix(pymongo): fix client validation issue (#8938)
Prior to this change, there were some scenarios in which pymongo would
call `validate_session` where it asserted that the client which created
the session was the client attempting to make a call with that session.

Since we are using ObjectProxy to wrap the MongoClient, we're actually
forwarding requests to a separate client. In this case, we run into
trouble in the validation situation where TracedMongoClient forwards
requests to MongoClient, which then creates a session and sets the
session create as itself (thus bypassing the proxy). After this fix, the
TracedMongoClient intercepts calls to create new sessions by MongoClient
and reassigns the client on the session to itself.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

(cherry picked from commit ee960a01675e3d9f22b070c51fec7d2753224348)