feat(llmobs): submit span events for chains from langchain (#8920)
## Summary
Follow-up to #8695 

This PR adds support for submitting LLMObs span events from the
LangChain integration automatically if LLMObs is enabled, and if the
LLMObs span is sampled. This is accomplished by:

- Setting the span type `SpanTypes.LLM` on `langchain` chain APM spans
so they are properly processed by the trace processor service in the
LLMObs service
- Tagging each chain span with additional _ml_obs.* tags, which get
popped from the trace when submitting the data they represent to LLMObs
intake through the LLMObs writer

## For Reviewers

Most of the LOC are test changes to account for the change in test
structure when dealing with variable amounts of expected calls to the
LLMObs writer. Additionally, many of the files changed are snapshot
files, to account for the change in the span type. Otherwise, the
important files changed are:

1. `ddtrace/contrib/langchain/patch.py`
2. `ddtrace/llmobs/_integrations/langchain.py`

Additionally, no release notes/changelog, as this is an internal change
for submitting span events to LLMObs intake.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)